correção bugs bandeira
