+++ .vscode/settings.json
@@ -42,6 +42,8 @@
         "compare": "cpp",
         "cstdint": "cpp",
         "matriz_led_control.h": "c",
-        "animacao_seta_jorge.h": "c"
-    }
+        "animacao_seta_jorge.h": "c",
+        "stdio.h": "c"
+    },
+    "C_Cpp.errorSquiggles": "disabled"
 }
