+++ RGB_LED_Matrix.c
@@ -46,6 +46,30 @@ pio_t meu_pio = {
     .b = 0.0, 
     .sm = 0          
 };
+void init_pio_routine(pio_t * meu_pio)
+{
+    //coloca a frequência de clock para 128 MHz, facilitando a divisão pelo clock
+    meu_pio->ok = set_sys_clock_khz(128000, false);
+
+    stdio_init_all();
+
+    printf("iniciando a transmissão PIO");
+    if (meu_pio->ok) printf("clock set to %ld\n", clock_get_hz(clk_sys));
+
+    //configurações da PIO
+    uint offset = pio_add_program(meu_pio->pio, &pio_matrix_program);
+    meu_pio->sm = pio_claim_unused_sm(meu_pio->pio, true);
+    pio_matrix_program_init(meu_pio->pio, meu_pio->sm, offset, OUT_PIN);
+}
+//rotina para definição da intensidade de cores do led
+uint32_t matrix_rgb(double b, double r, double g)
+{
+  unsigned char R, G, B;
+  R = r * 255;
+  G = g * 255;
+  B = b * 255;
+  return (R << 24) | (G << 16) | (B << 8)|0xFF;
+}
 
 void gpio_setup()
 {
@@ -161,7 +185,7 @@ void escolher_acao(char key)
 //função principal
 int main()
 { 
-    init_pio_routine(&meu_pio, OUT_PIN);
+    init_pio_routine(&meu_pio);
     
     // Configura os pinos do buzzer  Bcomo saída
     gpio_init(BUZZER_B_PIN);
@@ -187,15 +211,15 @@ int main()
 */
 while(true)
 {
- //   ligar_todos_os_leds_20_p(&meu_pio); 
- //   sleep_ms(300); 
- //   desliga_tudo(&meu_pio);
- //   sleep_ms(600); 
+    ligar_todos_os_leds_20_p(&meu_pio); 
+    sleep_ms(300); 
+    desliga_tudo(&meu_pio);
+    sleep_ms(600); 
     star_spangled_gleison(&meu_pio);
     
- //   sleep_ms(300); 
- //   desliga_tudo(&meu_pio);
-    //entrarModoBootloader();
+    sleep_ms(30000); 
+    desliga_tudo(&meu_pio);
+    entrarModoBootloader();
 }
     return 0;
 }
